Add iter_linear() and to_polystring() to QM
