+++ releasenotes/notes/polystring-d6ef3f4ec287cbac.yaml
@@ -4,3 +4,11 @@ features:
     Add ``BinaryQuadraticModel.iter_linear()`` method.
   - |
     Add ``BinaryQuadraticModel.to_polystring()`` method.
+  - |
+    Add ``QuadraticModel.iter_linear()`` method.
+  - |
+    Add ``QuadraticModel.to_polystring()`` method.
+upgrade:
+  - |
+    ``dimod.views.quadratic.QuadraticModelMixin`` now has an abstract property
+    ``.offset``.
